data_ingestion/pyproject.toml
[project]
name = "cloud-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.38.23",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
